Modify mapreduce to accomodate multiple jobs and distinct pipeline

+++ src/slave/hasher.rs
@@ -1,7 +1,7 @@
-pub fn get_hash(v: &str, lim: u8) -> u8 {
-    let mut sum: u8 = 0;
+pub fn get_hash(v: &str, lim: usize) -> usize {
+    let mut sum: usize = 0;
     for (i, l) in v.chars().enumerate() {
-        sum += (((i + 1) * (i + 1) * l as u8 as usize) % lim as usize) as u8;
+        sum += ((i + 1) * (i + 1) * l as u8 as usize) % lim;
     }
     return (sum % lim) + 1;
 }
